+++ src/data/series/IntSeries.java
@@ -4,6 +4,6 @@ package data.series;
  * Created by galafit on 22/9/17.
  */
 public interface IntSeries {
-    public int size();
-    public int get(int index);
+    public long size();
+    public int get(long index);
 }
