+++ src/data/series/StringSeries.java
@@ -4,6 +4,6 @@ package data.series;
  * Created by galafit on 26/9/17.
  */
 public interface StringSeries {
-    public int size();
-    public String get(int index);
+    public long size();
+    public String get(long index);
 }
