+++ src/data/series/grouping/aggregation/IntAggregateFunction.java
@@ -6,5 +6,5 @@ import data.series.IntSeries;
  * Created by galafit on 12/10/17.
  */
 public interface IntAggregateFunction {
-    public int group(IntSeries series, int from, int length);
+    public int group(IntSeries series, long from, int length);
 }
