+++ src/data/series/grouping/aggregation/IntMax.java
@@ -7,7 +7,7 @@ import data.series.IntSeries;
  */
 public class IntMax implements IntAggregateFunction {
     @Override
-    public int group(IntSeries series, int from, int length) {
+    public int group(IntSeries series, long from, int length) {
         int max = series.get(from);
         for (int i = from + 1; i < from + length; i++) {
             max = Math.max(max, series.get(i));
